Add Edge module to show alert if click on any link with # href Allow designers to put barriers in their prototype that don’t distract from user research.
For links to places not built yet, or places you want to remove access to, or external links, etc. You can use <a href="#"> and the user will see “Sorry this hasn’t been built yet” when they click it.

This works well in research, it’s clear what has happened, it means the link does not need to be removed and there is not a moment of confusion. The researcher can also prompt “What did you expect to see when you clicked that?”